fix(sdk): address review feedback
- remove name from DirectusExtension, update meta to match
ExtensionSettings
- remove registry-extension from schema in favor of inline registry
types
- remove DirectusExtensionRegistryList (use DirectusExtensionRegistry[]
instead)
- move reinstallRegistryExtension from create to update
- add readRegistryAccount and readRegistryExtension commands
- remove hardcoded output tests
- fix nit for alphabeitcal ordering
- update changeset details
