+++ .changeset/lucky-wolves-sing.md
@@ -2,4 +2,4 @@
 '@directus/sdk': minor
 ---
 
-Add SDK commands for the marketplace registry: `readRegistryExtensions`, `installRegistryExtension`, `reinstallRegistryExtension`, `uninstallRegistryExtension`
+Add SDK commands for the marketplace registry: `readRegistryExtensions`, `readRegistryAccount`, `readRegistryExtension`, `installRegistryExtension`, `reinstallRegistryExtension`, `uninstallRegistryExtension`
