+++ sdk/src/rest/commands/create/extensions.ts
@@ -20,21 +20,3 @@ export const installRegistryExtension =
 			method: 'POST',
 		};
 	};
-
-/**
- * Reinstall an extension from the registry.
- * @param extensionId - Registry extension ID
- * @returns Nothing
- * @throws Will throw if extensionId is empty
- */
-export const reinstallRegistryExtension =
-	<Schema>(extensionId: string): RestCommand<void, Schema> =>
-	() => {
-		throwIfEmpty(extensionId, 'Extension id cannot be empty');
-
-		return {
-			path: `/extensions/registry/reinstall`,
-			body: JSON.stringify({ extension: extensionId }),
-			method: 'POST',
-		};
-	};
