+++ sdk/src/rest/commands/create/index.ts
@@ -1,8 +1,8 @@
 export * from './collections.js';
-export * from './extensions.js';
 export * from './comments.js';
 export * from './dashboards.js';
 export * from './deployment.js';
+export * from './extensions.js';
 export * from './fields.js';
 export * from './files.js';
 export * from './flows.js';
