+++ sdk/src/rest/commands/update/extensions.ts
@@ -4,6 +4,24 @@ import type { RestCommand } from '../../types.js';
 import { throwIfEmpty } from '../../utils/index.js';
 
 /**
+ * Reinstall an extension from the registry.
+ * @param extensionId - Registry extension ID
+ * @returns Nothing
+ * @throws Will throw if extensionId is empty
+ */
+export const reinstallRegistryExtension =
+	<Schema>(extensionId: string): RestCommand<void, Schema> =>
+	() => {
+		throwIfEmpty(extensionId, 'Extension id cannot be empty');
+
+		return {
+			path: `/extensions/registry/reinstall`,
+			body: JSON.stringify({ extension: extensionId }),
+			method: 'POST',
+		};
+	};
+
+/**
  * Update an existing extension.
  * @param id - UUID of the extension
  * @param data - Partial extension object
