+++ sdk/src/schema/index.ts
@@ -6,7 +6,6 @@ export * from './core.js';
 export * from './dashboard.js';
 export * from './deployment.js';
 export * from './extension.js';
-export * from './registry-extension.js';
 export * from './field.js';
 export * from './file.js';
 export * from './flow.js';
